fix: float output, 2-format JSON match, fix auto-gen dtype issue

+++ op_builder/ops_src/pointnet2__ext_furthest_point_samplingSample/FrameworkLaunch/pointnet2__ext_furthest_point_sampling/op_host/pointnet2__ext_furthest_point_sampling.cpp
@@ -70,7 +70,7 @@ public:
             .UnknownShapeFormat({ge::FORMAT_ND});
         this->Output("y0")
             .ParamType(REQUIRED)
-            .DataType({ge::DT_INT32})
+            .DataType({ge::DT_FLOAT})
             .Format({ge::FORMAT_ND})
             .UnknownShapeFormat({ge::FORMAT_ND});
